pyproject.toml
[project]
name = "neuronest"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-generativeai>=0.8.5",
    "python-dotenv>=1.1.0",
    "streamlit>=1.46.0",
]
